msm: pil-q6v5-mss: Get smem reason buffer without taking remote spinlock
During a subsystem restart the modem might have crashed after taking the
remote spinlock. In such a case, when application processor tries to get
the smem entry for the SFR log, it would wait indefinitely for the remote
spinlock to be released. To solve this problem, use the smem API that
gets the entry buffer without taking the remote spinlock.

Change-Id: I00e77472bc88e276afb20835bdc24c5da09b9e86
Signed-off-by: Seemanta Dutta <seemanta@codeaurora.org>