gpio: msm: Remove software handling for dual-edge interrupts for TLMM_V2
The TLMM_V2 did not have hardware support for dual-edge interrupts.
This was handled in software using an algorithm to detect the current
edge and switch the interrupt polarity. This does not work after keeping
the RAW_STATUS_EN always enabled as switching the interrupt polarity
causes a spurious interrupt.
The TLMM_v3 has hardware support for dual-edge interrupts and this
continues to work as expected.
Thus, this patch removes dual-edge interrupt support for TLMM_V2.

Change-Id: I31651dea0b262adff3c3e0eede2ae83241dfe16b
CRs-Fixed: 520199
Signed-off-by: Rohit Vaswani <rvaswani@codeaurora.org>
Signed-off-by: houst0nn <houstonn@gmail.com>