input: touchscreen: Replace Kernel thread with a workqueue
Replace the Firmware update Kernel thread with a workqueue
instead in gt9xx_update.c Goodix driver. Now the firmware
upgrade procedure can be called later in future time without
blocking the registration of the touchscreen driver.
Using a kernel thread is an overhead in the current driver
as it is a one shot thread, hence replaced it with a workqueue.

Change-Id: I0d4731148351652092fe7feede0b44828939d98b
Signed-off-by: Shantanu Jain <shjain@codeaurora.org>
Signed-off-by: houst0nn <houstonn@gmail.com>