Revert "crypto: Fix memory leak in bam_release call"
This reverts commit I7aa13d5a03edbdfa806345039459feaeac6e7a11, which
release pbam memory when an error occurs. However, on a device with
multiple ce, pbam memory is shared by multiple pce_dev, it can only
be freed when the reference count goes to zero, and can not be
released it if reference count is not zero when an error occurs.
The previous implementation on bam_release was correct, so we
revert commit I7aa13d5a03edbdfa806345039459feaeac6e7a11.

Change-Id: Id8bab91f57fe8056e441afdeae45d7dabbd1eae8
Signed-off-by: Zhen Kong <zkong@codeaurora.org>
Signed-off-by: houst0nn <houstonn@gmail.com>