USB: ice40-hcd: Add ICE40 FPGA based SPI to Inter-Chip USB host driver
The ICE40 Host controller driver (ice40-hcd) is registered as a SPI
protocol driver and interacts with the SPI subsystem on one side and
interacts with the USB core on the other side. It supports only 1
Full-Speed port. The bridge chip does not maintain any port states.
Root hub port logic is simulated in the software. Control and Bulk
transfers are supported.

Change-Id: I927b4561a928f3eaa287351c4f82a5bef0cab76a
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: houst0nn <houstonn@gmail.com>